# backend/src/TaskManagement.Application/DTOs/TeamDtos.cs
using System.ComponentModel.DataAnnotations;

namespace TaskManagement.Application.DTOs;

public class TeamDto
{
    public Guid Id { get; set; }
    public string Name { get; set; } = string.Empty;
    public string? Description { get; set; }
    public Guid? ManagerId { get; set; }
    public string? ManagerName { get; set; }
    public int MemberCount { get; set; }
    public DateTime CreatedAt { get; set; }
    public List<UserDto> Members { get; set; } = new();
}

public class CreateTeamRequest
{
    [Required, MaxLength(150)]
    public string Name { get; set; } = string.Empty;

    [MaxLength(500)]
    public string? Description { get; set; }

    public Guid? ManagerId { get; set; }
}

public class UpdateTeamRequest
{
    [Required, MaxLength(150)]
    public string Name { get; set; } = string.Empty;

    [MaxLength(500)]
    public string? Description { get; set; }

    public Guid? ManagerId { get; set; }
}

public class AddTeamMemberRequest
{
    [Required]
    public Guid UserId { get; set; }
}